prometheus_grafana_fake_data/pyproject.toml
[project]
name = "monitoring-mcp-server"
version = "1.0.0"
description = "Production-grade MCP server for monitoring operations with fake data"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.2.0",
    "pyyaml>=6.0.1",
    "python-dateutil>=2.8.2",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
